Implemented GET /api/search/autocomplete endpoint
