+++ config/routes.yaml
@@ -1,13 +1,3 @@
-api_autocomplete_events:
-    path:       /api/autocomplete/events
-    controller: ChessApi\Controller\AutocompleteEventsController::index
-    methods:    GET
-
-api_autocomplete_players:
-    path:       /api/autocomplete/players
-    controller: ChessApi\Controller\AutocompletePlayersController::index
-    methods:    GET
-
 api_grandmaster:
     path:       /api/grandmaster
     controller: ChessApi\Controller\GrandmasterController::index
@@ -47,6 +37,11 @@ api_search:
     path:       /api/search
     controller: ChessApi\Controller\SearchController::index
     methods:    POST
+
+api_search_autocomplete:
+    path:       /api/search/autocomplete
+    controller: ChessApi\Controller\SearchAutocompleteController::index
+    methods:    GET
 
 api_player_stats:
     path:       /api/player_stats
