--- src/Controller/AutocompleteEventsController.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace ChessApi\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-
-class AutocompleteEventsController extends AbstractController
-{
-    const DATA_FOLDER = __DIR__.'/../../data';
-
-    const FILENAME = 'autocomplete-events.json';
-
-    public function index(Request $request): Response
-    {
-        $contents = file_get_contents(self::DATA_FOLDER.'/'.self::FILENAME);
-        
-        $response = new Response();
-        $response->setContent($contents);
-        $response->headers->set('Content-Type', 'application/json');
-
-        return $response;
-    }
-}
